Improvement of profile and booking schema
In this update the user is able to create an account, where it will be generated a token. If the user is logged in with the token, he will be able to create a profile. Once the user creates a profile an ID will be generated. So with the token and the profile ID the user is able to create a booking.